+++ app/accounts/urls.py
@@ -8,5 +8,6 @@ urlpatterns = [
 
     path('signup/', views.SignUpView.as_view(), name='signup'),
     path('activate/<uuid:username>/', views.UserActivateView.as_view(), name='user_activate'),
+    path('my-profile/', views.UserProfileView.as_view(), name='my_profile'),
 
 ]
